feat(Gyoungmin-min):[BOJ,17298] 오큰수 https://www.acmicpc.net/problem/17298
