+++ Gyoungmin-Min/baekjoon/src/baekjoon2/BOJ_3015.java
@@ -1,0 +1,2 @@
+package baekjoon2;public class BOJ_3015 {
+}
